Implement 'uvar' global
The work required to support `set` declarations among the globals
also required separating out the interpreter, making the globals
an object, and then wiring up interpreter<->globals to refer to
each other as equals. This is because

- any `set` declaration among the globals actually evaluates its
  expression and stores the result in the globals, so the globals
  depend on the interpreter

- (as before) the interpreter naturally refers to the globals

More precisely, as part of gradually building up the globals, the
globals need access to an interpreter which also has access to the
gradually built-up globals.

Technically, _all_ declarations among the globals require a
trip through the interpreter. But it just so happens that the ones
that are not `set` all evaluate `lit` literals, and so there the
trip is unnecessary.