Merge L::B::Globals into L::B::Interpreter
This was hinted at already since they were mutually dependent.

As part of this change, also turn all of the functions in the
interpreter into proper methods; since we're passing around the
interpreter (as `$self` or `$interpreter`) into all of the, we
no longer need to pass the `$s, $r, $m` triple. In fact, don't
worry about `$m` at all for now, since we don't have threads yet.

Having a tighter coupling between the interpreter and the globals
means we can make better use of intimate knowledge shared between
them. This will be exploited in the next commit.

At some point in the future, the globals will probably split again
from the interpreter, as we pre-compile all the definitions therein.
But for now, they belong together.