Make an improved version of `apply` form
This one avoids passing things into the evaluator in order to
build the arguments list. I've had a feeling for a while that
a simpler solution existed.

And it matters too; less for reasons of performance, and more
because the logger picked up a bunch of noise from that
artificial call. It pointed the way to this better solution.