Build the linked list at the same time as the hash
This slows down the test suite considerably; I think because the
order the globals are accessed is now predictable, and (maybe more
importantly) many of the commonly-called globals end up fairly
deep in the list.

Of course, with some searching we could find a better order for the
globals, to make the test suite run faster. But by the end of this
branch, it won't matter either way, because we'll be doing hash-based
lookups.