Move _id helper function Primitives->Types
It's not a primitive, and it belongs more properly to the helper
functions in Types.

Also took the chance to split it into five functions, because sometimes
we are able to call a narrowed variant from other parts of the
implementation.

- are_identical
    - pairs_are_identical
    - atoms_are_identical
        - chars_are_identical
        - symbols_are_identical

There should also be a `streams_are_identical` function too, in the
fullness of time.