Make sure 'smark' gets used in 'forms'
This was a subtle one; the serializer in the generator needed to
be told that `smark` is a little bit special and should be the one
to be generated into `forms`.

I don't know if this will happen with any other special pairs; it
didn't happen with `vmark` for example. If it does, we might want
to think up a less special-cased solution for it.