Allow clo literal without a body
This should be fine; the implementation was cranky about it because
it expected something where the body was supposed to be, but there
was nothing there.

All instances of pair_car() instead of prim_car() are gambles that
the call will always work, no matter what. If we're not sure, it's
far better to use the latter version.