Remove two no-longer-used test functions
These have been superseded by the DSL.

The `bel_todo` function remains used by one test file, but can go
when we no longer have any unimplemented features.

The `visit` function is of a slightly different nature; it's more of
a helper function to traverse directory structures. If it ends up
alone in this module, it might be high time to rename it.