Make sure to return when triggering error
This is good practice, as we don't want to continue running the
primitive at that point.

Note however that this has nothing to do with the _return value_
from the error function we're invoking. That one is mediated
through the interpreter's return value stack not through Perl's
return values. The `return` here is purely about control flow.