Fix failing dotted pair test
So, funny story.

`(1 . 2)`, while definitely a "dotted pair" syntactically, is
actually _not_ an improper list. The easiest way to see this is
that `2` is actually syntactic sugar for `(lit num ...)`, and
the outermost pair holding `lit` cancels out the dot, making
the list a proper list.

So that test actually failed (as it should). Rewriting it as
`(x . y)` (a properly improper list) fixed that.