Remove some fastfuncs
These are going because, as it turns out, they are _identical_
(as in, literally the same thing) to some other globals (which
all already have fastfuncs):

- i< is actually udrop
- srnum is actually cadr
- srden is actually caddr

In a later commit, we'll be binding these together directly,
which means the import of these (unused) fastfuncs would turn
into a test failure.

Interestingly, re-building the globals with the fastfuncs removed
already is less than perfectly straightforward, since (before the
globals have been rebuilt) the globals depended on those removed
fastfuncs, which are now import errors. Manual editing is
necessary for the globals to compile, so that they can properly
reproduce the same result. There might be a better way to do this,
but if there is, it escapes me right now.