Adjust tests
Now that the printer also prints labels, a surprising number of
the tests that apparently should contain labels started failing.
This commit corrects those tests, after manually confirming
that they should be corrected.

The test failures related to numbers were a little bit surprising.
But they are correct, because numbers are usually built by making
copies of `i1`; the one at the end is the actual pair `i1`.