Re-prioritize the test files
Based on the highly scientific method of throwing _all_ test files
out of the high-priority level, and then putting back only those
that clearly end up being last.

An exception to this is t/templates.t, which when putting it back
made the whole test suite slower.

There seems to be weak evidence that having too many tests
prioritized makes the whole test suite run slower, contrary to the
initial goal. As it is now, we have 5 files in 01 and 02, which
means that one thread will be free to get started on the faster
tests.