Implement 'nof' fast operative
Also implement fast operatives in general. The idea is similar to
fastfuncs, which are Perl versions of the Bel functions, running
faster both because there is no interpreter overhead and because
there is no overhead from lots of nested function calls which a
compiler could optimize away.

They are called fast operatives because like Kernel operatives and
unlike regular macros, they _run code_ rather than return an
expansion result for the interpreter to evaluate. Accordingly,
rather than interpolate the (code) arguments into an expansion
result, the operatives run their arguments in a given dynamic
environment -- an environment provided as an extra argument to the
operative.