Remove special-cased param opcodes
Replacing them instead with a single `param` opcode, which always
produces a list, possibly empty.

The reason for this change, besides having fewer and more
generalized opcodes, is to adjust for a future version of the
compiler that uses hypernets as an internal representation; it
was difficult to think of functions as taking an arbitrary number
of parameters in this future model, and much easier to think of
them as always accepting one parameter.